fixup! Move files from `protocol` to appropriate module
