+++ swap/src/asb.rs
@@ -1,16 +1,15 @@
-pub use rate::Rate;
-
-pub mod behaviour;
+mod behaviour;
 pub mod command;
 pub mod config;
-pub mod event_loop;
+mod event_loop;
 mod rate;
-pub mod recovery;
+mod recovery;
 pub mod tracing;
 pub mod transport;
 
 pub use behaviour::{Behaviour, OutEvent};
-pub use event_loop::{EventLoop, EventLoopHandle};
+pub use event_loop::{EventLoop, EventLoopHandle, FixedRate, KrakenRate, LatestRate};
+pub use rate::Rate;
 pub use recovery::cancel::cancel;
 pub use recovery::punish::punish;
 pub use recovery::redeem::{redeem, Finality};
