+++ swap/src/network/swap_setup.rs
@@ -1,10 +1,9 @@
+use crate::monero;
 use anyhow::{Context, Result};
 use libp2p::core::upgrade;
 use libp2p::swarm::NegotiatedSubstream;
 use serde::de::DeserializeOwned;
 use serde::{Deserialize, Serialize};
-
-use crate::monero;
 
 pub mod alice;
 pub mod bob;
