+++ swap/tests/alice_manually_redeems_after_enc_sig_learned.rs
@@ -3,8 +3,7 @@ pub mod harness;
 use harness::alice_run_until::is_encsig_learned;
 use harness::SlowCancelConfig;
 use swap::asb;
-use swap::asb::event_loop::FixedRate;
-use swap::asb::Finality;
+use swap::asb::{Finality, FixedRate};
 use swap::protocol::alice::AliceState;
 use swap::protocol::{alice, bob};
