+++ swap/tests/happy_path_restart_alice_after_xmr_locked.rs
@@ -2,7 +2,7 @@ pub mod harness;
 
 use harness::alice_run_until::is_xmr_lock_transaction_sent;
 use harness::SlowCancelConfig;
-use swap::asb::event_loop::FixedRate;
+use swap::asb::FixedRate;
 use swap::protocol::alice::AliceState;
 use swap::protocol::{alice, bob};
