+++ swap/tests/happy_path_restart_bob_after_xmr_locked.rs
@@ -2,7 +2,7 @@ pub mod harness;
 
 use harness::bob_run_until::is_xmr_locked;
 use harness::SlowCancelConfig;
-use swap::asb::event_loop::FixedRate;
+use swap::asb::FixedRate;
 use swap::protocol::bob::BobState;
 use swap::protocol::{alice, bob};
